+++ cmd/bnsd/app/app.go
@@ -15,7 +15,6 @@ import (
 	"github.com/iov-one/weave/cmd/bnsd/x/username"
 	"github.com/iov-one/weave/coin"
 	"github.com/iov-one/weave/commands/server"
-	"github.com/iov-one/weave/cron"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/migration"
 	"github.com/iov-one/weave/orm"
@@ -24,6 +23,7 @@ import (
 	"github.com/iov-one/weave/x/aswap"
 	"github.com/iov-one/weave/x/batch"
 	"github.com/iov-one/weave/x/cash"
+	"github.com/iov-one/weave/x/cron"
 	"github.com/iov-one/weave/x/currency"
 	"github.com/iov-one/weave/x/distribution"
 	"github.com/iov-one/weave/x/escrow"
