+++ x/cron/doc.go
@@ -1,0 +1,8 @@
+/*
+Package cron implements delayed message execution.
+
+This package provides a queue implementation for scheduling message for
+execution in the future and weave.Ticker compatible task runner.
+
+*/
+package cron
